Complete example added to readme (#8)
* Update README.md

* Update README.md

* Working example html file added  

This includes the example html file (supplied by Franck in response to an issue).  The previous examples are now in a subsection called More Examples.

* Added Node/Express server instructions

XMLHttpRequest can be a problem for development using the file system.  And since this library is geared towards local development rather than production, and it explicitly depends on XMLHttpRequest,  running a (any) web server (either local or remote) is a requirement.

* added webserver to requirements bullet list

* renamed file in example to index.html

renamed myFile.html to index.html in the example since index.html is the default file that will be served.  One less thing for new developers to worry about, and an easy change for anyone with experience.