Terminated Arrays
The arrays aren't terminated (before the elipses even), and it makes me twitch every time I see it.  I think an examples directory may be in order with fully functional, statically running examples.  I ran into a few cases where I really wanted a working example of the different ways that I could use the the http-vue-loader (which is just freaking awesome).  And I'm a big fan of both the most basic example (with names like my-component) along side examples that are still simple and focused (i.e. not using jquery or coffee script at the same time), but slightly more complex (and the names speak to solving a real problem... and the example actually does it).  I'll work on this.